Added a login component
Added a login component for users to login to their TMDB accounts and making it possible to see favorites list of the user logged in. I also changed the search results to keyword search rather than movie search so that anything that matches the searchterm can be seen as a result.